requirements.txt
Ghost.py == 0.2.3
beautifulsoup4 == 4.4.1
